Fix some packets

+++ src/command/gatt.rs
@@ -1306,7 +1306,7 @@ pub struct AddServiceParameters {
     /// The maximum number of attribute records that can be added to this service (including the
     /// service attribute, include attribute, characteristic attribute, characteristic value
     /// attribute and characteristic descriptor attribute).
-    pub max_attribute_records: usize,
+    pub max_attribute_records: u8,
 }
 
 impl AddServiceParameters {
